backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ccw-backend"
version = "0.1.0"
requires-python = ">=3.12,<3.14"
dependencies = [
"fastapi==0.111.0",
"uvicorn==0.29.0",
"pydantic==2.8.2",
"pydantic-settings==2.4.0",
"sqlalchemy==2.0.29",
"alembic==1.13.1",
"psycopg[binary]==3.1.18",
"redis==5.0.4",
"python-dateutil==2.9.0.post0"
]

[project.optional-dependencies]
dev = [
"pytest==8.2.2",
"pytest-cov==4.1.0",
"httpx==0.27.0",
"ruff==0.5.4",
"mypy==1.10.0",
"types-python-dateutil==2.9.0.20240316"
]
